refactor(auth): Move files
